backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
aiomysql==0.2.0
alembic==1.13.1

# Redis
redis==5.0.1
celery==5.3.6

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# LLM
litellm==1.35.0

# File processing
python-docx==1.1.0
pymupdf==1.23.8
python-magic==0.4.27

# Utils
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
requests
python-dotenv==1.0.0
bcrypt==4.1.2
cryptography==42.0.0
aiofiles==23.2.1

# Email (for notifications)
email-validator==2.1.0
